src/main/python/requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
mock==5.1.0
unstructured==0.10.25
typing_extensions==4.8.0
typing-inspect==0.9.0
langchain==0.0.320
google-cloud-storage==2.12.0
google-cloud==0.34.0
python-magic==0.4.24

